+++ assets/js/components/Sessions/CostHistoryChart.vue
@@ -179,7 +179,10 @@ export default defineComponent({
 						: this.$t("sessions.co2"),
 				data: Object.values(result).map((index) => index.avgCost),
 				yAxisID: "y1",
-				tension: 0.25,
+				tension: 0,
+				stepped: true,
+				borderJoinStyle: "round",
+				borderCapStyle: "round",
 				pointRadius: 0,
 				pointHoverRadius: 6,
 				borderColor: costColor,
